requirements.txt
pandas
matplotlib
reportlab
Pillow